requirements.txt
Django>=4.2.0
python-docx>=0.8.11
PyPDF2>=3.0.0
python-pptx>=0.6.21
reportlab>=4.0.0
python-magic>=0.4.27
clamd>=1.0.2
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
Pillow>=10.0.0
django-cleanup>=8.0.0
django-storages>=1.14.2
boto3>=1.34.0
django-environ>=0.11.2
pywin32>=306; sys_platform == 'win32' 